fixed regression in #2 + fixed other possible bug
So, it turns out #2 introduced a small regression when it comes to
resolving the path of the wallpaper set in Xresources. Now, one can use
variables and especially `~` again. I used `realpath -e` instead of
`eval echo` as `realpath` is the canonical solution for this problem
(looking at the man page: "Print the resolved absolute file name").

The other possible bug was at the end of the script, when it comes to
the GNOME terminal update routine. The command `xrescat gnome.terminal.update true`
would always return `true`, no matter the second key (i.e. I could write
`xrescat dawdmaldwmaldmkwamdkamdkaldw true` and it would return `true`).
I have therefore deleted the second `true` after the first argument of
the call to `xrescat`.

I'm sorry to have introduced the regression above. I thought I had
tested it well, but as Edsger W. Disktra said: "Program testing can
be used to show the presence of bugs, but never to show their absence!"