finally resolved the actual issue
So, `realpath` does not, against my expectations, resolve `~` in
strings, Bash does. I really want to avoid `eval` as it has some
inherent security flaws (the user can execute arbitrary code here if we
continue to use `eval`).
    Therefore, we will now manually replace `~` in the beginning with
the value of `${HOME}`, which is mimicking the behavior of Bash.
Then we run `realpath` to check the path. If the new file exists, we
can set it as the wallpaper. If the file path is set in `Xresources`,
but the path is invalid, we print a warning now, which is nice as it is
telling the user there is an error.

Moreover, I replaced `|| true` with `|| :` because it was late when I
wrote #2, and I completely forgot that the actual no-op in Bash is `:`.
This better communicates that nothing should be done in case the first
command fails.